requirements.txt
asgiref==3.8.1
audioread==3.0.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
Django==5.0.6
django-appconf==1.0.6
django-bootstrap3==24.2
django_compressor==4.5
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.60.0
numpy==2.0.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
soundfile==0.12.1
soxr==0.4.0
sqlparse==0.5.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
